feat(types): add shared relay DTO crate
