+++ Cargo.toml
@@ -4,6 +4,7 @@
 [workspace]
 members = [
     "crates/core",
+    "crates/types",
     "crates/adaptive",
     "crates/pii-redaction",
     "crates/cli",
@@ -24,6 +25,7 @@ repository = "https://github.com/NVIDIA/NeMo-Relay"
 
 [workspace.dependencies]
 nemo-relay = { version = "0.5.0", path = "crates/core", default-features = false }
+nemo-relay-types = { version = "0.5.0", path = "crates/types" }
 nemo-relay-adaptive = { version = "0.5.0", path = "crates/adaptive" }
 nemo-relay-pii-redaction = { version = "0.5.0", path = "crates/pii-redaction" }
 nemo-relay-ffi = { version = "0.5.0", path = "crates/ffi" }
