+++ crates/core/Cargo.toml
@@ -30,7 +30,7 @@ atof-streaming = [
     "tokio/rt-multi-thread",
     "tokio/time",
 ]
-schema = ["dep:schemars"]
+schema = ["dep:schemars", "nemo-relay-types/schema"]
 guardrails-remote = [
     "dep:reqwest",
     "dep:rustls",
@@ -75,6 +75,7 @@ openinference = [
 ]
 
 [dependencies]
+nemo-relay-types.workspace = true
 uuid = { workspace = true, features = ["v7", "serde"] }
 serde = { version = "1", features = ["derive", "rc"] }
 serde_json = "1"
