+++ crates/types/Cargo.toml
@@ -1,0 +1,26 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+[package]
+name = "nemo-relay-types"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "Shared serializable data model types for NeMo Relay."
+
+[lints]
+workspace = true
+
+[features]
+default = []
+schema = ["dep:schemars"]
+
+[dependencies]
+bitflags = { version = "2", features = ["serde"] }
+chrono = { version = "0.4", features = ["serde"] }
+schemars = { version = "0.8", optional = true }
+serde = { version = "1", features = ["derive", "rc"] }
+serde_json = "1"
+typed-builder = "0.23.2"
+uuid = { workspace = true, features = ["v7", "serde"] }
