+++ crates/types/src/api/mod.rs
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+//! Shared API data types.
+
+/// Lifecycle event DTOs.
+pub mod event;
+/// LLM DTOs and attributes.
+pub mod llm;
+/// Scope DTOs and attributes.
+pub mod scope;
+/// Tool DTOs and attributes.
+pub mod tool;
