+++ crates/types/src/codec/mod.rs
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+//! Shared normalized LLM request and response data types.
+
+/// Normalized LLM request data types.
+pub mod request;
+/// Normalized LLM response data types.
+pub mod response;
